+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+* The `Plugin.__call__` method now supports a `cleanup` argument
+  ([#92](https://github.com/watts-dev/watts/pull/92))
+
 ## [0.5.0]
 
 ### Added